+++ Classes/Attributes/Cache.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "feed_generator" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\FeedGenerator\Attributes;
+
+use Attribute;
+
+/**
+ * The attribute is used for the Cache-Control and Expires headers
+ * in the HTTP response of a feed.
+ */
+#[Attribute(Attribute::TARGET_CLASS)]
+final class Cache
+{
+    /**
+     * @param int<0, max> $seconds
+     */
+    public function __construct(
+        public readonly int $seconds,
+    ) {
+    }
+}