+++ Tests/Unit/Configuration/FeedConfigurationTest.php
@@ -1,0 +1,68 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "feed_generator" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\FeedGenerator\Tests\Unit\Configuration;
+
+use Brotkrueml\FeedGenerator\Configuration\FeedConfiguration;
+use Brotkrueml\FeedGenerator\Feed\FeedFormat;
+use Brotkrueml\FeedGenerator\Tests\Fixtures\FeedConfiguration\SomeFeed;
+use PHPUnit\Framework\TestCase;
+
+final class FeedConfigurationTest extends TestCase
+{
+    /**
+     * @test
+     * @doesNotPerformAssertions
+     */
+    public function instantiatingWithCacheInSecondsWithNullIsValid(): void
+    {
+        new FeedConfiguration(
+            new SomeFeed(),
+            '/some/path',
+            FeedFormat::ATOM,
+            [],
+            null,
+        );
+    }
+
+    /**
+     * @test
+     * @doesNotPerformAssertions
+     */
+    public function instantiatingWithCacheInSecondsWith0IsValid(): void
+    {
+        new FeedConfiguration(
+            new SomeFeed(),
+            '/some/path',
+            FeedFormat::ATOM,
+            [],
+            0,
+        );
+    }
+
+    /**
+     * @test
+     */
+    public function instantiatingWithCacheInSecondsWithANegativeNumberThrowsException(): void
+    {
+        $this->expectException(\DomainException::class);
+        $this->expectExceptionCode(1655707760);
+        $this->expectExceptionMessage('The configured cache seconds (-1) is a negative int');
+
+        new FeedConfiguration(
+            new SomeFeed(),
+            '/some/path',
+            FeedFormat::ATOM,
+            [],
+            -1,
+        );
+    }
+}